move chat to src
