+++ src/components/Address.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import { classNames } from '../helpers'
 import useLookup from '../hooks/useLookup'
