+++ src/index.ts
@@ -1,0 +1,7 @@
+import { XmtpProvider } from './components/XmtpProvider'
+import Layout from './components/Layout'
+
+export { Account } from './contexts/xmtp'
+
+export const Chat = Layout
+export const ChatProvider = XmtpProvider
