--- styles/Loader.module.css
@@ -1,84 +1,0 @@
-.lds-roller {
-  display: inline-block;
-  position: relative;
-  width: 40px;
-  height: 40px;
-}
-.lds-roller div {
-  animation: lds-roller 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
-  transform-origin: 20px 20px;
-}
-.lds-roller div:after {
-  content: ' ';
-  display: block;
-  position: absolute;
-  width: 6px;
-  height: 6px;
-  border-radius: 50%;
-  background: #31006e;
-  margin: 0px 0 0 0px;
-}
-.lds-roller div:nth-child(1) {
-  animation-delay: -0.036s;
-}
-.lds-roller div:nth-child(1):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(2) {
-  animation-delay: -0.08s;
-}
-.lds-roller div:nth-child(2):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(3) {
-  animation-delay: -0.12s;
-}
-.lds-roller div:nth-child(3):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(4) {
-  animation-delay: -0.16s;
-}
-.lds-roller div:nth-child(4):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(5) {
-  animation-delay: -0.2s;
-}
-.lds-roller div:nth-child(5):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(6) {
-  animation-delay: -0.24s;
-}
-.lds-roller div:nth-child(6):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(7) {
-  animation-delay: -0.28s;
-}
-.lds-roller div:nth-child(7):after {
-  top: 0px;
-  left: 25px;
-}
-.lds-roller div:nth-child(8) {
-  animation-delay: -0.32s;
-}
-.lds-roller div:nth-child(8):after {
-  top: 0px;
-  left: 25px;
-}
-@keyframes lds-roller {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
