--- styles/MessageComposer.module.css
@@ -1,32 +1,0 @@
-.bubble {
-  border-radius: 24px;
-  background-color: #fafafa;
-  border: 1px solid #e5e7eb;
-  height: 40px;
-  padding: 10px 8px 10px 16px;
-}
-
-.input {
-  font-size: 14px;
-  border: none;
-  background: none;
-  padding-left: 0;
-  margin-top: -2.8px;
-  height: 25px;
-}
-
-.input:focus {
-  outline: none;
-  border: none;
-  box-shadow: none;
-}
-
-.arrow {
-  /* 
-  // There must be a more flex-boxy way to do this.
-  // Problem is that this button needs to start higher than the text 
-  */
-  margin-top: -2.8px;
-  width: 25px;
-  height: 25px;
-}
