--- types/global.d.ts
@@ -1,19 +1,0 @@
-declare module 'react-blockies' {
-  import React from 'react'
-  interface BlockiesProps {
-    seed: string
-    size?: number
-    scale?: number
-    color?: string
-    bgColor?: string
-    spotColor?: string
-    className?: string
-  }
-  const Blockies: React.FC<BlockiesProps>
-
-  export default Blockies
-}
-
-interface Window {
-  ethereum: import('@metamask/providers').BaseProvider
-}
